Testing framework bootstrap
This commit introduces a series of primitive frameworks for testing. Using a series of command line flags, we can specify the manual IP's to launch each section of scalog on, so that we can run a local unit testing library. This is all very hacky right now, but take alook at the corresponding README.md in the testing directory for instructions on running the tests.

Future considerations: Allow users to specify IP's, startup and tear down data and order between end2end tests.