Created discovery command
This launches a new server for the express purpose of responding to clients with a list of available data servers that they can potentially talk to. After dockerizing, this should live as a deployment, where this discovery service is replicated behind the scenes, allowing for a reliable gateway to the Scalog service.