Finalization for scalog
This code touches the data and ordering layer. We assume that the kubernetes operator will be able to send a signal to the ordering layer, so the finalization process starts from there. Ordering should send a signal to the data layer to tell it to kill its communication channels and to stop taking append reqeusts from clients. Both layers should clean up metadata used in communication.