Update Raft paradigm to log contestedCuts instead of committedcuts
Prior to this commit, raft is utilized to achieve consensus across of a cluster of machines by running leader election and forwarding messages exclusively to that one leader. This implementation style, while functional, is difficult to predict and hard to maintain.

This commit restructures scalog's raft such that we achieve consensus on a log consisting of a series of contestedcuts sent to the ordering layer by the aggregators. The log of raft can now contain three types of messages -- cuts, batch commands, and finalization commands -- this allows our raft cluster to remain synchronized and propogate a consistent state across our data nodes.